Add support for compiling on Linux.
On windows the AST was converted straight to the PE format. This mean understanding and generating valid executables. To avoid this on Linux and speed the process up a file with AT&T style assembly is generated plus a small "C harness". The harness serves as the "proper" ELF main method which simply calls `clara_main` in the generated assembly.

Finally both files are linked together using `gcc` to produce the final valid ELF file.