Add support for function parameters.
All Node instances have been updated to contain a reference to the scope in which they are declared. This allows resolving identifier function call arguments to function parameters by the `VarSymbol`. This is symbol will link a function call identifier and the parameter it refers to. This is important as when generating function call code we need to know at which memory address we can find the argument to pass.

The `printDate` function can now be written in Clara from almost 2 years ago can now be implemented in Clara code! (for a fixed date value!)

In general the code is very ugly at present and almost certainly has bugs. The AST <-> Symbol relationship needs to be rationalised into something sensible. The SymTab is problematic in that it doesn't hand back instances of the symbol you want - just the Symbol interface.