Add support for basic `if` statements.
To do this the greater than operator (>) has also been implemented to implement the if condition expression. No other comparison operators are available. Currently the condition expression is limited to simple expressions consisting of only 1 operand on either side of the `>` symbol.

This is because operator precedence is not yet well defined and there is no support for parenthesis to capture precedence syntactically.