Add support dot selection of functions.
This allow users to extend any type with their "own" functions. All "dot selection" calls are simply an AST rewrite to pass the left operand of the dot select expression into the position 0 of the function call identified in the right operand.

The included file `dot-selection.clara` shows some examples of adding functions to bool and int types.