Add support for "simple" structs.
This enables simple aggregated types (int, bool & string). Structs containing structs are not yet supported.

By convention all struct identifiers must start with a lowercase letter because their constructor function follows the pattern of using the same name but with an an uppercase first letter. Example:

```
struct myStruct {
   s : string
}

fn main() {
   MyStruct("Hello") // Creates a new instance of `myStruct` with provided content
}
```

Constructor functions are auto-generated and cannot be overridden.

To support this `VarSymbol` addressing needed to change to distinguish between those on the stack (parameters only at this stage) and those in the heap (struct offsets at this stage). The symbol table should probably be reviewed after this as there is duplication and overlapping concerns in places.