Correct AST walk to actually be DFS.
While we are here rewrite the logic to assign declarations & types to identifiers to be much simpler. Now when typed variables are parsed (currently just function parameters) we immediately attempt to resolve their type. If it is already known - assign it. If not assign the type in the semantic pass to while resolving identifiers.