Add "type linking" for forward references.
This change allow Node.Syms to be updated with the named type as and when it becomes available during the parse. This simplifies the semantic analysis phase as before we get there are struct fields and method parameters and return types are guaranteed to have been resolved to some (possibly invalid) type.

This paves the way for proper type checking of function bodies.