Update strings so they understand and have access to their length.
This is made available through a "pseudo" field called "length". This is similar to Java arrays and indeed _our_ upcoming array implementation will support the same field.

The length is written into the first 8 bytes of the generated string literal and attempting to access the length simply performs a `movq` on the pointer location to get the value.

Unfortunately this does introduce a special case when printing strings using the libc `printf` function. This library expects a null terminated string. To provide it we must supply a increment the string pointer to point "past" the length to the data.