Add support for using arrays and treating strings as arrays of bytes.
This commit adds support for arrays of any basic type or struct, indexing the array and also allows strings to be treated as arrays of bytes. The `strings.clara` file demonstrates the usage with respect to strings.

There is, however, no way to create or populate arrays yet.

Still to be done:
 - Bounds checking when accessing the array elements.
 - Creating new empty arrays.
 - Defining array literals.