Add support for declaration and assignment operator `:=`.
This operator is used to declare a new variable and assign a value. Attempting to use the same operator on an already declared variable or parameter will result in an error. The type of the declared variable is defined by the result of type checking the expression on the right-hand side of the operator.

Updating existing variable values (assignment only) will be implemented later.