Enforce static linking via gcc.
It appears that Ubuntu 16.10 and above has been compiled with `-fPIE` (https://wiki.ubuntu.com/SecurityTeam/PIE). As the assembly code we generate is not Position Independent Code (PIC) the linker fails with:

```
 relocation R_X86_64_32 against '.rodata' can not be used when making a shared object; recompile with -fPIC
```

Fix this by statically linking the binary for the moment. Hopefully it should not be to difficult to update at some point in future.