Add support for "external" functions.
These are declared as function signatures without a body. The implementation is expected to be available at link time. The anticipation is that libc functions required to bootstrap the runtime will be added in this way and can therefore be using from other clara functions.

The added example shows using the libc `getcwd` function.