Add support for integer arrays.
Creation is implemented in C and is supported by a new `runtime.c` file which will provide support for data structures which cannot be represented in Clara yet and is automatically included during compilation.

The memory layout for string and integer arrays are identical in that an 8-byte length precedes the actual data.