Complete support for byte type.
Currently bytes can only be _read_ from string values using the index operator. This results in a value with type byte which can only be used in string writes (which shouldn't be allowed!) as mixing int & bytes together in arithmetic operations is not supported. On top of this bytes cannot be used in type declarations for functions or structs.

Remove these limitations by allowing byte types to be declared everywhere and also automatically support casting from int to byte when necessary to allow mixing types without introducing a new operator.

To support mixing byte & int together all arithmetic operations now generate int values regardless of the types used and are automatically cast to byte if the destination requires it (local var, return type, parameter arg). It should be noted that this means that even byte + byte results in int. See `numbers.clara` for a full set of examples.

Byte values are stored on the stack & in structs using 8-bytes slots. This is simpler from a code generation point of view as all fields and local vars have the same width and also requires less assembly to generate the casts.