Add support for byte arrays.
This is implemented via the C runtime code just as for int arrays. To support for the `movb` instruction has been added to support writing a single byte into the array. To support reading a single _signed_ byte the previous implementation which performed `& 0xFF` on the result register has been replaced with correct usage of the `movsbq` instruction to read and sign extend a single byte to a quad word.