Correct string literal length calculation during ASM generation.
Currently the length of the string was determined by simply removing the double quotes from the start and end of the string. This is insufficient when the string contains escape characters such as `\n` as this will be considered as a string of two characters; `\` and `n`.

Fix this _unquoting_ string (i.e. convert all escape sequences back into their "raw" form) and measure this instead.