Correct type width calculation.
Currently this attempts to give the memory layout size of the type when actually it should report the number of bytes required to store a _reference of that particular type_.

Update it to do this and introduce a size method on `StructType` to report the size of a struct in memory. Also introduce a constant `ptrSize` for use when calculating stack slots.