Perform 2 type checking passes over AST.
The first pass type checks all function signatures and the second pass then type checks the function bodies. This separation allows the type checking code to guarantee that all function symbols will have type information for their parameters & defaults (if any).

This change simplifies the current code a little but is really to lay the groundwork for full function overloading in a later commit.