Implement function overloading.
Currently functions names must be globally unique in the user program and standard library. This limitation is problematic because it makes sense to have multiple functions named "append" which operate on different types.

Remove this limitation by allowing only functions to have multiple symbols which map to the same name. When this happens a linked list structure is used to chain the symbols together (much like collisions in a hash map). In addition a "marker" symbol is stored with the type information as part of the name which allows for easy detection and reporting of redeclared functions.

For instance the following function:

```
fn func1(i: int, b: byte, s: string) { }
```

will result in the addition of two symbols with the following keys:

```
func1
func1(int, byte, string)
```