Ensure symbol "redeclare" checks only examine current scope.
Previously many checks for existing symbols during parse attempted to resolve identifiers against the _entire symbol table_ hierarchy rather than just the current scope. This resulted in the situation where a function named `size` would effectively shadow any variable or parameter with the same name.

Fix this by updating Define(*Symbol) to return the symbol supplied or the existing symbol in conjunction with a boolean flag to indicate whether the a symbol with that name existed _in the current scope_. Using this method we can remove calls to Resolve() & simplify the parsing code.