Allow any number of function calls or index operators to follow operand.
Currently the parser will not recognise an index operator which immediately follows a function call.

Correct this by allowing any number of index operators and/or function calls to follow an identifier. This allows for immediate indexing of a function call return value (such as a string) & multidimensional arrays (not current supported).

This also paves the way for allowing a function to return a function which is then immediately called. The changes required to permit passing functions as parameters and returning functions from functions will come later.