Add support for "first-class" functions.
This commit allows functions to be passed as parameters to functions, returned from functions and stored in parameters for later invocation. Anonymous functions and closures are not yet supported. See `fn-as-params.clara` for a full set of examples.

The main changes required to support this are:

 * New syntax to specify a function as a parameter or return type.
 * Assembler output changes to support indirect calls.

The main limitation to this implementation is that invoking functions stored in structs does not work as the AST rewrite for dot-selection attempts to find a global function with the field name. For example:

```
struct s {
  f: fn(int) int
}

fn double(i: int) int { return i * i }

fn main() {
  S(double).f(2) // ERROR: Cannot resolve function 'f(s, int)'
}

```

A future commit will correct the AST rewrite to dereference the function before the call.