Correct codegen for string & array types when calling external functions.
Both of these types are pointers to a block of memory which contains an 8-byte length followed by the actual data. Currently only string types have their pointer correctly incremented by 8 when passing the value to an external function.

Fix by also doing the same for arrays, update the C runtime `toString()` function to account for this and also ensure that these rules apply to _any_ external function - not just `printf()` as happened previously.