Add support for "while" statements.
This commits implements support for a basic C/Java while loop which has the following form:

```
 while bool-expr {
   ... statements ...
 }
```

It does not add support for `break` or `continue` statements. These will be added as and when they are required.

The original plan for looping was to use default parameters and tail recursion. While still the long term goal, I'm unsure how to implement them at present & it would help the standard library to have a looping construct now.

As and when default parameters & tail recursion become available the `while` statement can be removed from the language.