Refactor malloc calls to be made from Clara via 'claralloc' function.
This commit ensures that all memory allocation happens from a single place in Clara code, namely the `claralloc` function which returns a `pointer` struct. This is a stepping stone to adding GC to the language.

As a side-benefit the `runtime.c` code has now been simplified to simply enabling casts. All code for creating arrays and copying strings has been moved into Clara.