Expose nothing type for use in programs.
This is necessary to resolve ambiguity when defining external functions which return nothing from those that may return a function. Consider the following example:

```
fn ext1() // Returns nothing
fn ext2() int

```
The parser current will see the `fn` keyword of `ext2` as the start of a `fn()` return type of `ext1`. This is not a problem with normal function definitions as they match on either a type of a starting brace and can detect `nothing` return type as such.

This could be fixed by allowing the parser to see newline characters but that is a much bigger change. This should be revisited in future however.