Support runtime access & navigation of the frame pointer.
All function calls in Clara maintain a so called "frame pointer". This value points to the stack location of the _calling_ function. _That_ stack location yet again stores the stack location of the caller of _that_ function and so on. This forms a linked list all the way up to the stack base.

This commit does 2 things:

1) It sets the frame pointer of the C `main` function as the stack base.
2) It provides Clara functions to get & navigate the frame pointer.

Currently nothing useful can be done with this functionality but we will soon be adding GC code which will need the ability to determine & use the return address of each frame on the stack.