Correct AST rewrite when calling fns stored in structs.
Currently the rules for dot selection mean that when a function call is detected on the right of a '.' operator the AST is rewritten to pass the left node as the first parameter. This allows for easy chaining of functions.

However in the case that the function being called is stored in a struct (as functions are first class citizens) there are two problems:

1) the left is passed as a parameter when it should not.
2) the field symbol name is output as the function name in the generated assembly.

This commit fixes both by detecting when the left node is a struct and the right node is a function call whose name matches a struct field and instead rewrites the AST to calculate the address of the field and _then_ perform an indirect call on the result.