Support runtime log messages via builtin `debug` function.
The intention is that this function will enable debug output of specific parts of the runtime via commandline flags in later commits. This functionality is most relevant to the forthcoming GC code.

The intended usage is as follows:

```
debug("malloc", "- Allocated %d bytes.", size)

```
The C runtime would then check if the `malloc` log had been enabled and if so output the log line to the console.