Fix instruction selection for byte array writes.
When writing a value to a byte array the compiler should emit a `movb` instruction. The current code attempts to do this by examining the left AST node of an assignment statement. However the left type holds the array _element_ type and the information that the left is actually from an array is lost during typecheck. As such the generated code assumes that the byte value is actually stored on the stack and as such has an 8-byte width.

The end result of this was that storing values in arrays worked but also scribbled zeroes over the 7 bytes _past end of the array_ which is used by libc to track the memory it has returned from calls to `malloc`. The final result was heap corruption.

Fix this by ensuring the symbol on the left of an `opArray` node is promoted to the parent node and use that during code generation to check if the left node is a byte array.