Correct parameter symbol construction for constructor functions.
Currently the symbols which are defined for struct fields are reused for the generated constructor function parameters. This is an issue because the field offsets in those symbols are overwritten during code generation by the assignment of stack slot offsets as part of the standard function calling convention.

This is currently not a problem because the original field offsets are not actually needed during code generation. However the forthcoming garbage collector needs these to generate tracing function calls for all stack slots which are pointers into the heap.

As such fix this by copying the struct field AST node & associated symbol.