Add support for garbage collection.
This commit adds a precise, non-moving, incremental mark & sweep GC to the language. It implements the tag-free approach outlined here: https://cs.nyu.edu/goldberg/pubs/gold91.pdf to provide the GC "plumbing" which mark & sweep is layered on top of. It also adds a `debugGc` flag to the C runtime to enable verbose logging of both the mark & sweep steps. Crucially the paper also details how to support closures & polymorphic types both of which are planned for inclusion in the language.

To support tracing GC roots & walking the heap calls to malloc allocate an extra 16 bytes. The first 8 bytes is used to maintain a linked list of all pointers returned by malloc. The second 8 bytes is used as an "administrative header" which tracks whether the pointer was traced during marking and if the pointer is readonly (this is the case for string literals).

Numerous improvements are possible to the current implementation. First, the sweeping stage requires clearing the marked bit. This means performing a write to every object in the heap. Instead a counter can be maintained in the header, incremented during the mark and compared to the GC iteration number. Also clearing the GC mark & checking the mark does not use bitwise operations as it should because the language does not yet support them.

References:

[1] Tag-Free Garbage Collection for Strongly Typed Languages: https://cs.nyu.edu/goldberg/pubs/gold91.pdf

[2] Modern Compiler Design: Chapter 5, Section 2

[3] Baby's First Garbage Collector: http://journal.stuffwithstuff.com/2013/12/08/babys-first-garbage-collector/