Fix variable scoping error due to eager symbol declaration.
Currently a symbol is defined when parsing a declaration and assignment statement. The new variable name may shadow a global reference of the same name. This cane be problematic if a statement _earlier_ in the function expects to resolve the global identifier rather than the function local identifier. This happens because the symbol table is only consulted for identifiers during typecheck.

The following example illustrates the problem:

```
fn x() {
  y()    // Invoke global "y" function
  y := z
  y()    // Invoke global "z" function via variable
}

fn y() { printf("Y") }
fn z() { printf("Z") }
```
This program should print "YZ" but instead prints "ZZ" as the first y function call has been resolved to the identifier y on the next line.

Fix this by delaying variable symbol definition until after the right hand side of the `:=` has been typechecked.

In future the parsing, symbol table construction & type-checking phases should be rewritten to be cleanly separated from each other.