Rewrite parser to separate AST & symbol table construction.
Currently symbol table creation and population is largely handled during the parse. This leads to significant tangling of the recursive descent parser and symbol table management. It also leads to a confusing "2 pass" typecheck due partial handling of forward declarations while building the AST.

Fix this by rewriting the parser to completely separate AST construction from symbol table scoping & population. The most significant changes from this are that the AST now properly represents type information & the "2 pass" type check is no longer required as the initial population of the symbol table is carefully done to not require it.

The original idea had been to separate the front-end into 3 completely separate parts:
 -- parse: AST construction only
 -- symtab build & use: Build symbol table and use it to find all identifiers
 -- typecheck: ensure all assigned symbols & types are contextually valid

Unfortunately due to the simple "type inference" provided by the `:=` operator, I believe symbol creation & typechecking must be mixed together.

Consider the following example:
```
x := <expr>
z := x.y
```
Resolving symbol y requires that we know the _type_ of x as y could be a global function or a field of x.