Correct type matching for numeric arrays.
Due to automatic casting support between byte and int values their array counterparts were also accidentally interchangeable. This could cause a byte array to be passed to a function expecting an int array which. This would result in a read or write of 8-byte values instead of 1 byte and vice versa in the generated code.

Fix this by ensuring when matching an array type we match the element type _exactly_ in the case that it is a byte or int.