Cleanup ASM formatting and support descriptions for instructions.
Replace tabs with 3 spaces and use padding to ensure the first operand to all instructions starts in the same place vertically. Also add the ability to output a description of what the instruction is doing if debug mode is enabled for the code generator.

At present no descriptions have been added. The upcoming rewrite of the code generator will make use of this.