Add support for end-to-end (e2e) test cases.
Update the existing test harness to load & execute "test" Clara programs. These programs are normal Clara which support adding "structured comments" to indicate expected lines of output when the program is executed. A structured comment is simply a comment with the form:

```
// EXPECT: <...output...>
```

A test file adds these at the end of `printf` statement like follows:

```
printf("Hello!\n") // EXPECT: Hello!
```

The test harness parses all comments, executes the test, parses the standard output from the run and checks that the produced output matches what was expected.

Other changes in this commit simplify the compiler entry point and ensure that all logging is written to a provided `io.Writer`. This allows the test harness to `/dev/null` these writes and not clutter the test output.