Ensure debug log messages are flushed immediately to standard out.
This ensures that when debugging a segfaulting program all log messages to just before the crash make it to standard out. This is particularly useful when debugging GC errors. Also update the e2e test harness to output any captured standard out messages when there is a failure to execute the generated binary.