Use 64-bit moves when referring to external symbols.
Currently Clang produces the following error when attempting to compile compiler generated assembly code:

```
error: 32-bit absolute addressing is not supported in 64-bit mode
```

This is because the referenced symbol could be at an offset greater than 32-bits which cannot be encoded in a normal ^`movq` instruction. To allow for a full 64-bits the `movabs` instruction must be used instead.