Ensure stack is always 16-byte aligned.
Currently the code generator makes no attempt to ensure that $rsp is always aligned on a 16-byte boundary. This directly contradicts the AMD 64-bit ABI which states:

```
The end of the input argument area shall be aligned on a 16 (32, if __m256 is passed on stack) byte boundary.
```

(Particularly puzzling is that on Linux this seemed to cause no issues. I have not determined why.)

Fix this by ensuring that all stack manipulation leaves $rsp as a multiple of 16.