Correct linker symbol name generation & referencing on OSX.
OSX requires that all global linker symbols be prefixed with an underscore. This historically appears to have been a workaround to avoid name clashes with existing assembly libraries (see here for more info: https://stackoverflow.com/a/5908708)

As such update all symbol names on OSX to have and use the same scheme.