Fully support 64-bit integer literals.
Currently the compiler always generates `movq` instructions when loading integer literals. However the max length encoding of this instruction cannot represent all 8-bytes of a 64-bit literal. Also the compiler does not check that the literal can be represented in 8-bytes.

Fix both these problems by parsing the literal to ensure it can be represented in 8-bytes and also updating the code generator to output `movabs` instructions for literals requiring more than 32-bits of storage.

This commit also introduces a simple form of constant folding to ensure that the two node AST structure which represents a negative integer literal is collapsed into one before attempting to check for overflow.