Add support for anonymous functions & closures.
Anonymous functions, that is functions that only reference variables local to that function (or globals), are implemented by simple "function hoisting". This involves an AST rewrite to move the function definition to the root node, giving it a name & simply referencing that name at the previous declaration site.

Closures on the other hand are a much more complicated beast...

Implementing closures involves 2 major AST modifications:

 1. Transforming the closure declaration site into a "closure" struct which captures all required state.
 2. Transforming _all_ non-global function call sites to ensure the referenced function is _correctly_ invoked.

Step 1 requires the following:

 1(a) Identify all free variables & auto-generate an "environment" struct with fields for them.
 1(b) Hoist the function declaration to the root node & give it a new name.
 1(c) Rewrite the function to take the newly generated "environment" struct as its first parameter and then rewrite all free variable references to go "through" the "env" struct.
 1(d). Auto-generate a "closure" struct which holds a reference to the now top-level closure function & the "environment" for that function.

 For example a program which looks like this:

 ```
 fn main() {
    f := x(10)
    f(20)
 }

 fn x(i: int) fn() int {
    return fn(j: int) int = i + j
 }
 ```

 Would be transformed into the following code:

 ```
 fn main() {
   f := x(10)
   f()
 }

 fn x(i: int) fn() int {
   env := Environment(i)
   closure := Cl(clFn.3AA12BF, env)
   return closure
 }

 fn clFn.3AA12BF(environment env, j: int) int = env.i + j

 struct cl {
   func: fn() int
   env: environment
 }

 struct environment {
   i int
 }
 ```

Step 2 requires generating the appropriate "call sequence" for the referenced function.

To continue from the example above, invoking the closure 'f' would look like the following:

```
 fn main() {
   f := x(10)
   f.func(f.env, 20)
 }

 // ... as before ...
```

The issue with this is that at compile time there is no way to know whether a variable or struct field which holds a function contains a reference to a "normal" function or a closure. This must be dynamically detected at runtime by some sort of probing.

To do this a new assembly-defined function called 'invokeDynamic' has been added. It is applied to all non-global function call sites via an AST rewrite and accepts a reference to the function plus all its arguments. It determines how to invoke the supplied function by checking for the "read-only" flag being set in the GC header. If this is the case we have a "normal" function. If not we have a closure and it is invoked like the code example above. This requires that the assembly generator output a "read-only" GC header before all function declarations just like for string literals.

Therefore the example from above would now be as follows:

```
 fn main() {
   f := x(10)
   invokeDynamic(f, 20)
 }

 // ... as before ...

```

The other major change is providing GC support to trace closures at runtime. The basic idea is to output the address of a GC tracing function for the closure struct _before_ the declaration of the closure function. Detecting a closure is performed the same as for 'invokeDynamic'. The `Tag-Free Garbage Collection for Strongly Typed Languages` paper (https://cs.nyu.edu/goldberg/pubs/gold91.pdf) provides a full description of the approach.

References:

[1] Modern Compiler Design: 6.3.6 Lambda Lifting

[2] How To Compile Lambda: http://matt.might.net/articles/closure-conversion/