Provide accurate start positions for function args & index operators.
Currently the identifier `lex.Token` is reused when building AST nodes for chained function calls or index operators. This means that position information for any errors concerning these nodes is always wrong.

Fix this by returning both the AST node for the function args or index operator plus the '(' or '[' token to provide accurate position information.