Refactor kind-specific function information.
Currently the `FunctionType` struct uses boolean flags to indicate the "kind" of the function. These are normal, external, constructor & closure. These flags are also supplemented with other fields which are only set for specific kinds (e.g. "gcFunc" for `closure`). This is already confusing and is set to get worse with the upcoming enum implementation.

Fix this by adding a `FuncKind` type to represent the different kinds and factor out kind-specific information into a "Data" field of type `interface{}`.