Add support for enums.
This commit introduces support for tagged unions (more loosely enums) as a new top-level type. Enums allow the creation of a type which may be one of a few different so called "variants".

The following example illustrates an example declaration:

```
enum notification {
   Email(subject: string, from: string, message: string)
   Sms(from: string, message: string)
   Voicemail(from: string, durInSecs: int, message: []byte)
}

fn main() {
   n1 := Email("Hello!", "some@body.com", "Hi there, I'm writing to...")
   n2 := Sms("+4477123456789", "Yo!")
   n3 := Voicemail("+4477987654321", 120, byteArray(100))

   // n1, n2 & n3 all have type 'notification'
}

```

At present the language provides no way access the details of a particular variant. A future commit will add the 'match' statement which will declare 'case's for each member to allow access to variant-specific data.