Add middle & backend support for AST blocks.
A block is a lexically scoped collection of zero of more statements within curly braces. For example:

```
fn main() {
   x := 1
   {
      y := 1
      z := y + x
   }
}

```

It should be noted that this commit does not provide _front-end support_ for parsing the above code. The intention is that the upcoming `match` keyword and associated semantics will generate blocks during AST rewrites.