Support 'matching' over enums.
Currently there is no way to understand which member of an enum is stored in a variable. This commit fixes that by adding support for 'matching' semantics by use of the 'match' statement. The following example highlights its use:

```
fn main() {

    e := WindowResized(10, 20, 150, 200)

    match e {
        case Refresh():
            printf("refreshed!)\n")

        case KeyPressed(key):
            printf("pressed key = %s\n", key)

        case WindowResized(x, y, w, h):
            area := w * h
            printf("new window area = %d\n", area)

        case MouseDown(x, y):
            printf("clicked @ (%d, %d)\n", x, y)
    }
}

enum windowEvent {
    Refresh()
    KeyPressed(key: string)
    MouseDown(x: int, y: int)
    WindowResized(x: int, y: int, w: int, h: int)
}

```

As can be seen match statements accept an expression and contain a series of zero or more case blocks. When inside a case block programs have de-structured access to enum member fields. At present match statement can only be used by enums.

It should be noted that there are no changes to the code generator to support these new keywords. Case statement variables are rewritten as normal variable declarations & the match statement itself is lowered to an if/elseif block.

At present matching is not exhaustive but the intention is that a future commit will make it so.