Ensure $rsp is 16-byte aligned before all CALL instructions
The System V AMD64 ABI, 3.2.2 The Stack Frame, states that:

```
The end of the input argument area shall be aligned on a 16 (32 or 64, if __m256 or __m512 is passed on stack) byte boundary. In other words, the value(%rsp + 8)is always a multiple of 16(32 or 64) when control is transferred to the function entry point. The stack pointer, %rsp, always points to the end of the latest allocated stack frame
```

The current code attempts to do this by inserting specific alignment instructions at various points in the generated code. Unfortunately this is insufficient and leads to panics when calling libc function such as printf.

Fix this by tracking the current alignment for each function as it is generated. Then use this information when generating CALL instructions to ensure that the $rsp alignment is always correct, modifying it before & after the CALL if necessary.