Rewrite various C runtime functions in Clara
Currently, functions to abort a running program when an index out of bounds error occurs or a program invokes the panic function are defined in C. There is no need for this as they can be expressed in Clara code instead. As such remove the C code.

This change is part of a series of commits to move as much code from C & assembly into Clara as possible.