Only log heap addresses which have been marked
Currently the GC debug logging during the marking phase always pointer type it inspects regardless of whether it will actually be marked or not. This leads to a confusing log where it appears that string literals (stored in the read-only section of the binary) are being marked by the GC.

Correct this by skipping logging pointer types which are read-only or have already been marked.