Log when an entire frame is skipped during the GC mark phase
Currently frames which require no tracing when walking the stack invoke the `noGc()` function in the Clara runtime. This means that no frame information is printed by GC debug logging. This can be confusing to read as it appears frames are missing.

Address this by simply logging that that frame did not require any tracing. A future commit might find a way to output the frame name without generating lots of largely empty GC trace functions.