Add runtime support for unsafe heap operations
This commit adds a new function to the runtime library to support arbitrary pointer arithmetic & casting from <-> to any type. Its implementation & call site usage are completely excluded from typechecking to support the lack of an "any" type in the compiler.

It represents a total type system escape hatch & as such is extremely dangerous.

At present it is unused but it is intended to support the upcoming GC refactoring to allow dynamic lookup of GC tracing functions for heap objects from type information which will be stored in the block header value.

Hopefully future compiler work will find a way to severely limit its power or indeed remove it altogether. At the very least a future "any" type plus parameterised types would allow for valid typechecking.