Check & clear only the GC marked bit during sweep
Currently this code uses equality & assignment operators to manage the GC bit which is sufficient at present. However upcoming GC changes will store a GC type ID in the top 16 bits of this header which must be preserved throughout the lifetime of the heap allocation.

As such use the newly introduced bitwise operators to check & clear _only_ this bit in the header.