Encode type ID in block header on heap allocation
This commit ensures that all heap allocations encode a numeric type ID in the most significant 16-bytes of the block. Most ids are assigned during the generation of constructor functions although some are fixed for builtin types.

This information will be used by the upcoming GC refactor to relieve the compiler of having to generate per-stackframe GC functions.