Add GC root maps to support GC refactoring
Currently a bespoke function is generated for tracing the stackframe at each point where GC can occur (currently, function call sites). These generated functions must know the exact types of each root so that the correct GC function can be invoked. When polymorphic type parameters are added it will no longer be possible to know the exact types at compile and this scheme won't work.

This commit lays another piece of groundwork for moving away needing to know the exact types at compile type & supporting this information at runtime instead.

It does this by also generating GC root map for each GC tracing function. A root map is an array of offsets from the frame pointer which are known to point into the heap. This will be walked by the upcoming GC refactor to understand which frame slots to trace.