Refactor GC mark phase to use GC root maps to trace the frame
This commit refactors the GC mark phase to use the previously added root maps in conjunction with the type info table to dynamically trace GC roots in stack frames. It does this by outputting the address of the root map after a function call _instead_ of a bespoke function which previously did the tracing.

The GC then uses the root map info in conjunction with the type info table to:

 1. Get a pointer to the heap block
 2. Check the block header for the read only flag & skip if set
 3. Lookup a GC tracing function based on an id from the block header.
 4. Invoke the function to trace the heap block.

This logic means that compile time type information of stack slots is no longer required & a block of assembly code emitted at compile time can be removed. This will be done in a later commit.