Rework GC debug output to provide visual structure when tracing types
Currently all marked heap allocations appear in a flat list when debug logging is enabled. This doesn't show any relationship between the allocations, for instance if an allocation has been traced from walking a "parent" allocation.

This commit changes that and adds a simple indentation scheme to show this information.